requirements.txt
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.0
streamlit>=1.22.0
plotly>=5.14.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
polars>=0.19.15
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
schedule>=1.2.0